sync github code to coding

+++ build/build.sh
@@ -29,16 +29,17 @@ done
 
 if [ $use_zfp == "y" ] && [ ! -n "${ZFP_HOME}" ]; then
   export ZFP_HOME=/usr/local/include/
-  rm -rf zfp*
-  wget ${ZFP_URL} -O zfp.tar.gz
-  tar -xzvf zfp.tar.gz
-  pushd zfp* 
-  mkdir build && cd build
-  cmake ..
-  cmake --build . --config Release
-  make install
-  popd
-  rm -rf zfp*
+  if [ ! -d $ZFP_HOME ]; then
+    rm -rf zfp*
+    wget ${ZFP_URL} -O zfp.tar.gz
+    tar -xzf zfp.tar.gz
+    pushd zfp-0.5.5
+    mkdir build && cd build
+    cmake ..
+    cmake --build . --config Release
+    make install
+    popd
+  fi
 fi
 
 OS_NAME=$(uname)
@@ -51,12 +52,11 @@ else
     if [ ! -d "${ROCKSDB_HOME}" ]; then
       rm -rf rocksdb*
       wget  ${ROCKSDB_URL} -O rocksdb.tar.gz
-      tar -xzvf rocksdb.tar.gz
-      pushd rocksdb*
+      tar -xzf rocksdb.tar.gz
+      pushd rocksdb-6.2.2
       CFLAGS="-O3 -fPIC" make shared_lib -j
       make install
       popd
-      rm -rf rocksdb*
     fi
   fi
 fi
@@ -68,7 +68,7 @@ echo "version info: $flags"
 echo "build gamma"
 pushd $GAMMAOUT
 cmake -DPERFORMANCE_TESTING=ON -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$ROOT/ps/engine/gammacb/lib $ROOT/engine/
-make && make install
+make -j && make install
 popd
 flatbuffers=$ROOT/ps/engine/third_party/flatbuffers-1.11.0
 if [ -d ${flatbuffers} ]; then
@@ -78,8 +78,25 @@ fi
 export LD_LIBRARY_PATH=$LD_LIBRARY_PATH:$ROOT/ps/engine/gammacb/lib/lib/
 export LIBRARY_PATH=$LIBRARY_PATH:$ROOT/ps/engine/gammacb/lib/lib/
 
-echo "build vearch"
-go build -a -tags="vector" -ldflags "$flags" -o $BUILDOUT/vearch $ROOT/startup.go
+function build_vearch(){
+  echo "build vearch"
+  if [ $1 == 'mod' ];then
+    echo "build vearch by mod"
+    export GO111MODULE="on"
+    go build -mod=mod -a -tags="vector" -ldflags "$flags" -o $BUILDOUT/vearch $ROOT/startup.go
+    echo "build deploy tool by mod"
+    go build -mod=mod -a -ldflags "$flags" -o $BUILDOUT/batch_deployment $ROOT/tools/deployment/batch_deployment.go
+  else 
+    echo "build vearch by vendor"
+    export GO111MODULE="off"
+    go build -a -tags="vector" -ldflags "$flags" -o $BUILDOUT/vearch $ROOT/startup.go
+    echo "build deploy tool by vendor"
+    go build -a -ldflags "$flags" -o $BUILDOUT/batch_deployment $ROOT/tools/deployment/batch_deployment.go
+  fi
+}
 
-echo "build deploy tool"
-go build -a -ldflags "$flags" -o $BUILDOUT/batch_deployment $ROOT/tools/deployment/batch_deployment.go
+if [ $# == 1 ];then
+  build_vearch 'mod'
+else
+  build_vearch 'vendor'
+fi
